Back-end stock creation and logic
Added to the stock and user controllers and requests, as well as models to make the many to many relationship between them work.
Added a button to the dashboard view to test this, currently logs the users' own stocks.
userID is now also saved inside the session storage for simplicity