Portfolio management
Added get and post logic to the stock controller, validation rules etc. on the back-end. Front-end also has the needed functions to interface with the portfolios.
Changed some of the entry types on the migrations for parity between variable names
Using all of the above, the currently showing stock details can now be stored to your portfolio and fetched on the dashboard.