Portfolio view, Stock controller/DB improvements
Added a portfolio view/component on the front-end that now houses the userstock/portfolio api fetches.
A lot of work done on the back-end to prevent duplicate entries in both stocks and M2M relationships.
Added industry column to Stock model
Made the API output a lot more similar to that of finnhub to make reusing components more viable.