Stocklinechart, configurable candle data fetch
Finished and cleaned up the linechart, now stocklinechart.
Added a second data source in the chart with own axis etc.
Added a way of making the asked time segment of price data variable.
Aligned the pagecomponent body in the center of the view instead of left while retaining it's max width