Dashboard news, selectmenu, portfolio
Added a news section to the dashboard with for now a selection from a few categories, comes from the store and pagination/search easily to implement.
Made the selectmenu component fully reusable, added it to the dashboard to filter by category.
Added a little overview part to the portfolio page with the number of portfolio entries for the current user