Added stockinfo implementation
Implemented logic and added to the migration for the 'shareprice' variable in order to calculate portfolio value down the line.
Added 'amount' variable to the pivot table to track the number of a certain stock the user ownes, the stockresource now also returns this column when fired.
When storing a new stock now check whether the pivot table should be either updated or extended.
Front-end now also sends along the amount of a certain stock when filled in by the user.